Change the definition of 'dpm deps'
The definition of 'dpm deps' has been changed.  Previously it showed the
dependency tree whose root is PKGNAME; now it shows any dependency tree
that includes PKGNAME.  PKGNAME is highlighted in red when output to a
terminal.

The behavior when no argument is specified remains unchanged, showing
the dependency tree of all direct installed packages.